refactor: abstract EmojiPicker component
