+++ src/styles/components/_KeyboardAccessory.js
@@ -1,8 +1,5 @@
 import { StyleSheet } from 'react-native';
 import colors from '../common/_colors';
-import Dimensions from 'Dimensions';
-
-const window = Dimensions.get('window');
 
 export default StyleSheet.create({
   container: {
@@ -21,15 +18,5 @@ export default StyleSheet.create({
   item: {
     color: '#979797',
     width: 40
-  },
-  image: {
-    height: 50,
-    width: (window.width - 10) / 7,
-    padding: 5
-  },
-  pageView: {
-    flex: 1,
-    flexDirection: 'row',
-    flexWrap: 'wrap',
   }
 });
